+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Change Log
 ==========
 
+ * build artifacts are reported to console
+
 Version 0.14.0 (2017.11.20)
 ---------------------------